WorldMap moved into organisms
