+++ resources/js/Components/organisms/home/WorldMap/constants/MapPoints.ts
